Moved sqlalchemy dependant code from common.py
